BC-2306 move DelayController to timeout.interceptor.spec.ts
