+++ apps/server/src/shared/common/interceptor/timeout.interceptor.spec.ts
@@ -1,7 +1,39 @@
-import { HttpStatus, INestApplication } from '@nestjs/common';
-import { TimeoutInterceptor } from '@shared/common';
+import { Controller, Get, HttpStatus, INestApplication, NestInterceptor } from '@nestjs/common';
+import { APP_INTERCEPTOR } from '@nestjs/core';
+import { Test, TestingModule } from '@nestjs/testing';
+import { RequestTimeout, TimeoutInterceptor } from '@shared/common';
 import request from 'supertest';
-import { createTestModule } from './test/create-test.module';
+
+const delay = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
+
+@Controller()
+class DelayController {
+	/** default route to test public access */
+	@Get()
+	async getHello(): Promise<string> {
+		await delay(100);
+		return 'Schulcloud Server API';
+	}
+
+	@RequestTimeout(1)
+	@Get('/timeout')
+	async getHelloWithTimeout(): Promise<string> {
+		await delay(100);
+		return 'Schulcloud Server API';
+	}
+}
+
+export const createTestModule = (interceptor: NestInterceptor): Promise<TestingModule> => {
+	return Test.createTestingModule({
+		providers: [
+			{
+				provide: APP_INTERCEPTOR,
+				useValue: interceptor,
+			},
+		],
+		controllers: [DelayController],
+	}).compile();
+};
 
 describe('TimeoutInterceptor', () => {
 	describe('when integrate TimeoutInterceptor', () => {
